fix(RLXStatus): ensure accurate TPS calculation by explicitly casting to double
- Updated TPS calculation in RLXStatus to explicitly cast timeSinceLastTpsCalculation to double, improving precision in the calculation.
- This change enhances the reliability of TPS metrics displayed in the client sidebar.

+++ src/plugin/RLXStatus.cpp
@@ -41,7 +41,7 @@ void RLXStatus::oneTick() {
     auto timeSinceLastTpsCalculation =
         std::chrono::duration_cast<std::chrono::seconds>(now - lastTpsCalculationTime_).count();
     if (timeSinceLastTpsCalculation >= 1) {
-        double tps = static_cast<double>(tickCount_) / timeSinceLastTpsCalculation;
+        double tps = static_cast<double>(tickCount_) / static_cast<double>(timeSinceLastTpsCalculation);
 
         auto level = ll::service::getLevel();
         if (!level.has_value()) return;
